Cico 38994

+++ rover/controllers/rateManager_/rvRateManagerLeftSideFilterCtrl.js
@@ -85,15 +85,36 @@ angular.module('sntRover').controller('rvRateManagerLeftSideFilterCtrl', [
 
         if (!!selectedRateType && !alreadyExistInSelectedRateTypeList) {
           $scope.selectedRateTypes.push(selectedRateType);
+        }
+
+        $scope.deleteAllSelectedCards();
+      };
+
+      /**
+       * Rate type selected from rate types tab.
+       * on choosing the rate type from list, we will be adding to selected list
+       */
+      $scope.rateTypeSelectedFromRTT = () => {
+        if ($scope.selectedRateTypeIDFromRTT.trim === '') {
+          return;
+        }
+
+        var conditionToTest = {id: parseInt($scope.selectedRateTypeIDFromRTT)},
+
+          selectedRateTypeFromRTT = _.findWhere($scope.rateTypes , conditionToTest),
+
+          alreadyExistInSelectedRateTypeListFromRTT = (_.findIndex($scope.selectedRateTypesFromRTT, conditionToTest) > -1);
+
+        if (!!selectedRateTypeFromRTT && !alreadyExistInSelectedRateTypeListFromRTT) {
+          $scope.selectedRateTypesFromRTT.push(selectedRateTypeFromRTT);
 
           //adding the elements will change the height
           refreshScroller();
 
           //setting the focus to newly added rate type
-          scrollTo('#selected-rate-type-list span:last-child');
+          scrollTo('#selected-rate-type-list br');
         }
 
-        clearAllRatesAndAllRoomTypes();
         $scope.deleteAllSelectedCards();
       };
 
@@ -112,21 +133,33 @@ angular.module('sntRover').controller('rvRateManagerLeftSideFilterCtrl', [
       };
 
       /**
+       * to delete a rate type selected, from Rate Type Tab
+       * @param  {LongInteger} rateTypeID [selected rate type's id to delete]
+       */
+      $scope.deleteSelectedRateTypeFromRTT = (rateTypeID) => {
+        var indexToDelete = _.findIndex($scope.selectedRateTypesFromRTT , {id: parseInt(rateTypeID)});
+        $scope.selectedRateTypesFromRTT.splice(indexToDelete, 1);
+
+        $scope.selectedRateTypeIDFromRTT = '';
+        
+        //deleting the node will change the height
+        refreshScroller();
+      };
+
+      /**
        * to remove all selected rate type in one take
        */
       $scope.deleteAllSelectedRateTypes = () => {
         $scope.selectedRateTypes = [];
         $scope.selectedRateTypeID = '';
-        //deleting the nodes will change the height
-        refreshScroller();
       };
 
       /**
-       * utility function to clean the ALL RATES/ALL ROOM TYPE radio box
+       * to remove all selected rate type in one take from Rate Type Tab
        */
-      var clearAllRatesAndAllRoomTypes = () => {
-        $scope.showAllRates = false;
-        $scope.showAllRoomTypes = false;
+      $scope.deleteAllSelectedRateTypesFromRTT = () => {
+        $scope.selectedRateTypesFromRTT = [];
+        $scope.selectedRateTypeIDFromRTT = '';
       };
 
       /**
@@ -159,10 +192,9 @@ angular.module('sntRover').controller('rvRateManagerLeftSideFilterCtrl', [
             refreshScroller();
 
             //setting the focus to newly added rate
-            scrollTo('#selected-rate-list span:last-child');
+            scrollTo('#selected-rate-list br');
           }
 
-          clearAllRatesAndAllRoomTypes();
           $scope.deleteAllSelectedCards();
         }
       };
@@ -187,8 +219,8 @@ angular.module('sntRover').controller('rvRateManagerLeftSideFilterCtrl', [
       $scope.deleteAllSelectedRates = () => {
         $scope.selectedRates = [];
         $scope.selectedRateID = '';
-        
-        // deleting the nodes will change the height
+
+        //deleting the node will change the height
         refreshScroller();
       };
 
@@ -203,36 +235,17 @@ angular.module('sntRover').controller('rvRateManagerLeftSideFilterCtrl', [
           return true;
         }
 
-        var selectedRateTypeIDs = _.pluck($scope.selectedRateTypes, 'id');
-
-        return (selectedRateTypeIDs.indexOf(rate.rate_type.id) > -1);
+        return (parseInt($scope.selectedRateTypeID) === rate.rate_type.id);
       };
 
       /**
-       * on tapping the ALL RATES tab
+       * clear all selected values on changing tab
        */
-      $scope.changedAllRatesSelection = () => {
-        // we will clear out all selected from other tab
-        $scope.deleteAllSelectedRates();
-        $scope.deleteAllSelectedRateTypes();
-      };
 
-      /**
-       * on tapping the ALL RATE TYPES tab
-       */
-      $scope.changedAllRateTypes = () => {
-        // we will clear out all selected from other tab
+      $scope.deleteAllSelectedValues = () => {
         $scope.deleteAllSelectedRates();
         $scope.deleteAllSelectedRateTypes();
-      };
-
-      /**
-       * on tapping the ALL ROOM TYPES tab
-       */
-      $scope.changedAllRoomTypes = () => {
-        // we will clear out all selected from other tab
-        $scope.deleteAllSelectedRates();
-        $scope.deleteAllSelectedRateTypes();
+        $scope.deleteAllSelectedRateTypesFromRTT();
         $scope.deleteAllSelectedCards();
       };
 
@@ -263,21 +276,22 @@ angular.module('sntRover').controller('rvRateManagerLeftSideFilterCtrl', [
               $scope.showAllRates = true;
               $scope.showAllRateTypes = false;
               $scope.showAllRoomTypes = false;
-              $scope.changedAllRatesSelection();
+              $scope.deleteAllSelectedRateTypesFromRTT();
               break;
 
         case 'RATE_TYPES' : 
               $scope.showAllRates = false;
               $scope.showAllRateTypes = true;
               $scope.showAllRoomTypes = false;
-              $scope.changedAllRateTypes();
+              $scope.deleteAllSelectedRates()
+              $scope.deleteAllSelectedRateTypes();
               break;
 
         case 'ROOM_TYPES' : 
               $scope.showAllRates = false;
               $scope.showAllRateTypes = false;
               $scope.showAllRoomTypes = true;
-              $scope.changedAllRoomTypes();
+              $scope.deleteAllSelectedValues();
               break;
       }
 
@@ -289,16 +303,30 @@ angular.module('sntRover').controller('rvRateManagerLeftSideFilterCtrl', [
       switch($scope.chosenTab) {
 
         case 'RATES' : 
-              buttonText = 'Show All Rates';
+              if ($scope.selectedRates.length === 0 && $scope.selectedCards.length === 0) {
+                buttonText = 'Show All Rates';
+              }
+              else if ($scope.selectedCards.length > 0 ) {
+                buttonText = 'Show Contract Rates';
+              }
+              else {
+                buttonText = 'Show Selected Rates';
+              }
               break;
 
         case 'RATE_TYPES' : 
-              buttonText = 'Show All Rate Types';
+              if ($scope.selectedRateTypesFromRTT.length === 0 ) {
+                buttonText = 'Show All Rate Types';
+              }
+              else {
+                buttonText = 'Show Selected Rate Types';
+              }
               break;
 
         case 'ROOM_TYPES' : 
               buttonText = 'Show All Room Types';
               break;
+
         default :
               buttonText = 'Show All Rates';
       }
@@ -338,10 +366,13 @@ angular.module('sntRover').controller('rvRateManagerLeftSideFilterCtrl', [
       /**
        * utility method to clear all selection from tabs
        */
-      const clearAllRatesAllRoomTypesAllRateTypesAllRates = () => {
-        clearAllRatesAndAllRoomTypes();
-        $scope.deleteAllSelectedRateTypes();
+      const clearAllRatesAllRoomTypesAllRateTypes = () => {
         $scope.deleteAllSelectedRates();
+        $scope.deleteAllSelectedRateTypes();
+        $scope.deleteAllSelectedRateTypesFromRTT();
+
+        //deleting the node will change the height
+        refreshScroller();        
       }
 
       /**
@@ -358,7 +389,7 @@ angular.module('sntRover').controller('rvRateManagerLeftSideFilterCtrl', [
           }
         }
 
-        clearAllRatesAllRoomTypesAllRateTypesAllRates();
+        clearAllRatesAllRoomTypesAllRateTypes();
         
         $scope.showAllRates = true;
         
@@ -500,7 +531,14 @@ angular.module('sntRover').controller('rvRateManagerLeftSideFilterCtrl', [
        * This method handles on-click of the SHOW RATES BUTTON
        */
       $scope.clickedOnShowRates = () => {
+        var selectedRateTypeList;
         //PAGINATION stuff will be handled from RateManagerCtrl
+        if ($scope.chosenTab === 'RATES') {
+          selectedRateTypeList = $scope.selectedRateTypes;
+        }
+        if ($scope.chosenTab === 'RATE_TYPES') {
+          selectedRateTypeList = $scope.selectedRateTypesFromRTT;
+        }
         var valuesChoosed = {
           fromDate: $scope.fromDate,
           toDate: $scope.toDate,
@@ -509,11 +547,13 @@ angular.module('sntRover').controller('rvRateManagerLeftSideFilterCtrl', [
 
           groupBy: $scope.groupBySelectedValue,
 
+          chosenTab: $scope.chosenTab,
+
           showAllRates: $scope.showAllRates,
           showAllRateTypes: $scope.showAllRateTypes,
           showAllRoomTypes: $scope.showAllRoomTypes,
 
-          selectedRateTypes: $scope.selectedRateTypes,
+          selectedRateTypes: selectedRateTypeList,
           selectedRates: $scope.selectedRates,
 
           selectedCards: $scope.selectedCards,
@@ -556,6 +596,10 @@ angular.module('sntRover').controller('rvRateManagerLeftSideFilterCtrl', [
         $scope.rateTypes = []; //will be filled from API once we get to th = view
         $scope.selectedRateTypes = [];
         $scope.selectedRateTypeID = ''; //ng-model for rate type selection
+
+        //rate type from Rate Type Tab
+        $scope.selectedRateTypeIDFromRTT = ''; //ng-model for rate type selection
+        $scope.selectedRateTypesFromRTT = [];
 
         //rate related
         $scope.rates = []; //will be filled from API once we get to th = view
